move ec add to accelerated
