+++ p256/Cargo.toml
@@ -29,7 +29,7 @@ sha2 = { version = "0.10", optional = true, default-features = false }
 [target.'cfg(all(target_os = "zkvm", target_arch = "riscv32"))'.dependencies]
 risc0-bigint2 = { git = "https://github.com/risc0/risc0", rev = "e41742c5af32222b1ae3ffbea8fb9a8b1181ab64", features = ["unstable"] }
 bytemuck = "1"
-# risc0-bigint2 = { version = "1.2.0", features = ["unstable"] }
+# risc0-bigint2 = { version = "~1.2.0", features = ["unstable"] }
 
 [dev-dependencies]
 blobby = "0.3"
